semester2/project/requirements.txt
# numpy
# numpy-quaternion
PyGLM
matplotlib
